JENA-2837: Reduce use of model APIs in response handling
