+++ jena-fuseki2/jena-fuseki-access/src/main/java/org/apache/jena/fuseki/access/SecurityContextView.java
@@ -27,10 +27,9 @@ import java.util.function.Predicate;
 import org.apache.jena.graph.Node;
 import org.apache.jena.query.Query;
 import org.apache.jena.query.QueryExecution;
-import org.apache.jena.query.QueryExecutionFactory;
-import org.apache.jena.query.QueryFactory;
 import org.apache.jena.sparql.core.DatasetGraph;
 import org.apache.jena.sparql.core.Quad;
+import org.apache.jena.sparql.exec.QueryExec;
 import org.apache.jena.sparql.util.Context;
 import org.apache.jena.sparql.util.NodeUtils;
 import org.apache.jena.tdb1.TDB1Factory;
@@ -87,20 +86,15 @@ public class SecurityContextView implements SecurityContext {
     }
 
     @Override
-    public QueryExecution createQueryExecution(String queryString, DatasetGraph dsg) {
-        return createQueryExecution(QueryFactory.create(queryString), dsg);
-    }
-
-    @Override
-    public QueryExecution createQueryExecution(Query query, DatasetGraph dsg) {
+    public QueryExec createQueryExec(Query query, DatasetGraph dsg) {
         if ( isAccessControlledTDB(dsg) ) {
-            QueryExecution qExec = QueryExecutionFactory.create(query, dsg);
+            QueryExec qExec = QueryExec.dataset(dsg).query(query).build();
             filterTDB(dsg, qExec);
             return qExec;
         }
         // Not TDB - do by selecting graphs.
         DatasetGraph dsgA = DataAccessCtl.filteredDataset(dsg, this);
-        return QueryExecutionFactory.create(query, dsgA);
+        return QueryExec.dataset(dsgA).query(query).build();
     }
 
     /**
@@ -108,7 +102,7 @@ public class SecurityContextView implements SecurityContext {
      * {@link QueryExecution}. This does not modify the {@link DatasetGraph}.
      */
     @Override
-    public void filterTDB(DatasetGraph dsg, QueryExecution qExec) {
+    public void filterTDB(DatasetGraph dsg, QueryExec qExec) {
         GraphFilter<?> predicate = predicate(dsg);
         qExec.getContext().set(predicate.getContextKey(), predicate);
     }
